Delete code gigantic security hole fixed
Before this fix, it was possible to send a POST to /edit/delete and
basically delete the whole database as there was no authorization. 

The work of a sys admin never stops.

+++ src/main/java/com/codepump/service/impl/CodeServiceImpl.java
@@ -77,8 +77,9 @@ public class CodeServiceImpl implements CodeService {
 	}
 
 	@Override
-	public void deleteCode(int id) {
-		dbServ.deleteCodeItem(id);
+	public void deleteCode(int id, String SID) {
+		if (authServ.authoriseDelete(id, SID))
+			dbServ.deleteCodeItem(id);
 
 	}
